rearrange files, add exceptions
