+++ models/initializer.rb
@@ -1,0 +1,3 @@
+require_relative 'person'
+require_relative 'bankAccount'
+require_relative 'transaction'
